requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.6
django-leaflet==0.27.1
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
